requirements.txt
requests
pandas
mysql-connector-python